Renaming SpecDiff Package to Diff and MOving to the cmd/registry package.
