+++ cmd/registry/diff/test-specs/base-test.yaml
