+++ cmd/registry/diff/test-specs/struct-test-add.yaml
