+++ cmd/registry/diff/test-specs/struct-test-delete.yaml
