+++ cmd/registry/diff/test-specs/struct-test-modify.yaml
