+++ cmd/registry/diff/test-specs/struct-test-order.yaml
